tolerancia como double. agregada view segmentada para variaciond e parametros

+++ esteban/gtp07/e03/p01.cpp
@@ -7,10 +7,10 @@ int main( int argc, char **argv ) {
     const char *filename = cimg_option( "-f",
                                         "../../imagenes/bone.tif",
                                          "ruta archivo imagen" );
-    int x_inicial  = cimg_option ( "-x" ,   200, "valor x inicial" );
-    int y_inicial  = cimg_option ( "-y" ,   200, "valor y inicial" );
-    int tolerancia = cimg_option ( "-tol" , 40.0, "tolerancia en intencidad" );
-    int vecinos    = cimg_option ( "-v" , 4, "cantidad vecinos (4-8)" );
+    int x_inicial     = cimg_option ( "-x" ,   200, "valor x inicial" );
+    int y_inicial     = cimg_option ( "-y" ,   200, "valor y inicial" );
+    double tolerancia = cimg_option ( "-tol" , 40.0, "tolerancia en intencidad" );
+    int vecinos       = cimg_option ( "-v" , 4, "cantidad vecinos (4-8)" );
 
     CImgDisplay disp, disp1, disp2, disp3, disp4, disp5, disp6;
